+++ src/DataProtection/Changelog.md
@@ -18,7 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* General availability of `Az.DataProtection`
+* Fixed list parameter set for `Get-AzDataProtectionBackupVault`
 
 ## Version 0.5.0
 * Added support for automatic assignment of permissions for configure backup for DPP workloads - AzureDisk, AzureBlob, AzureDatabaseForPostgreSQL using Set-AzDataProtectionMSIPermission cmdlet